+++ src/output/output_entity.rs
@@ -1,9 +1,5 @@
 use anyhow::{anyhow, Error, Result};
-use serde::Serialize;
-use std::{
-    convert::TryFrom,
-    fmt::{self, Display},
-};
+use std::{convert::TryFrom, fmt};
 
 /// Represents the available output formats.
 #[derive(Debug, PartialEq)]
@@ -34,7 +30,7 @@ impl TryFrom<Option<&str>> for OutputFmt {
     }
 }
 
-impl Display for OutputFmt {
+impl fmt::Display for OutputFmt {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         let fmt = match *self {
             OutputFmt::Json => "JSON",
@@ -45,12 +41,12 @@ impl Display for OutputFmt {
 }
 
 /// Defines a struct-wrapper to provide a JSON output.
-#[derive(Debug, Serialize, Clone)]
-pub struct OutputJson<T: Serialize> {
+#[derive(Debug, Clone, serde::Serialize)]
+pub struct OutputJson<T: serde::Serialize> {
     response: T,
 }
 
-impl<T: Serialize> OutputJson<T> {
+impl<T: serde::Serialize> OutputJson<T> {
     pub fn new(response: T) -> Self {
         Self { response }
     }