+++ src/smtp/mod.rs
@@ -1,0 +1,1 @@
+//! Module related to SMTP.